Changes: - UI Functions have been sorted out - New use of forward declarations for classes (didn't know you could do that)
Issues:
- For some reason, it isn't loading resources (failing to find the files, probably a directory structure issue again)

+++ GameProject/Game.cpp
@@ -71,7 +71,7 @@ using namespace std;
 
 		player.position += thrust * elapsed;
 		thrust = Decay(thrust, 0.1f, 0.1f, elapsed);
-		Vector2f gravity = { 0, Defaults::gravityFactor };
+		Vector2f gravity = { 0.f, (float)Defaults::gravityFactor };
 		thrust += gravity;
 
 		if (player.position.y < (player.bounds.height * 0.1f))
